Refactor codebase and apply multiple bugfixes
This commit includes a major refactoring of the application, splitting the monolithic app.py into several focused modules for better organization and maintainability:
- app.py (main entry point)
- main_window.py (MainWindow class)
- accordion_widget.py (AccordionWidget class)
- settings_window.py (SettingsWindow class)
- ping_worker.py (PingWorker class)
- resize_grip.py (ResizeGrip class)

Additionally, numerous bug fixes and improvements were implemented:
- Enhanced error handling for settings.json loading and saving.
- More robust loading of game and map icons, with better error logging and use of placeholders.
- Strengthened network request error handling, including timeouts and user alerts for critical API failures.
- Implemented file-based logging to app.log (DEBUG level).
- Fixed and enabled frameless window resizing using a custom resize grip.
- Improved diagnostic logging for PingWorker failures.
- Corrected API URL construction for fetching server details.
- Addressed issues with window width and height persistence, ensuring immediate updates after settings changes and smoother accordion transitions.
- Resolved Matplotlib warnings by converting data to numerical types for plotting.
- Ensured settings dialog checkboxes correctly reflect current game enabled/disabled states.
